+++ application/frontend/config/main.php
@@ -44,11 +44,12 @@ return [
                 /*
                  * User routes
                  */
-                'GET  user'                           => 'user/index',
-                'GET  user/<employeeId:\w+>'          => 'user/view',
-                'POST user'                           => 'user/create',
-                'PUT  user/<employeeId:\w+>'          => 'user/update',
-                'PUT  user/<employeeId:\w+>/password' => 'user/update-password',
+                'GET  user'                                  => 'user/index',
+                'GET  user/<employeeId:\w+>'                 => 'user/view',
+                'POST user'                                  => 'user/create',
+                'PUT  user/<employeeId:\w+>'                 => 'user/update',
+                'PUT  user/<employeeId:\w+>/password'        => 'user/update-password',
+                'PUT  user/<employeeId:\w+>/password/assess' => 'user/assess-password',
 
                 /*
                  * Authentication routes